Made separate classes for each of the data sources and used them in the audit script
